libcst: Hide node's children by default.
LibCST tries to make CST tree uniform and that is great for traversal,
but unfortunately adds a lot of clutter when trying to visualize it.

For most purposes it is enough to see which nodes are contained in the
attributes, so we hide children of the node by default. There is still
an option to display them by passing option "--show-children".